requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# osubg.py: 4
Pillow == 6.1.0

# main.py: 9,10
PyQt5_sip == 12.8.0

# osu_audio_gen.py: 2
numpy == 1.19.3

# osu_audio_gen.py: 4
pydub == 0.25.1

# dbparse.py: 2
tqdm == 4.45.0
